EmuThread: add way to specify update interval and display fps as float instead of int for better accuracy

+++ src/core/core.h
@@ -16,7 +16,7 @@ enum class State {
     Idle,
 };
 
-using UpdateFunction = std::function<void(int fps)>;
+using UpdateFunction = std::function<void(float fps)>;
 
 class Core {
 public:
